feat(votepeloclima): refactor read_csv_file function
